type harmonization

+++ src/render/axis_cross.cpp
@@ -5,19 +5,19 @@
 #include <vector>
 
 constexpr double pi{3.14159265359};
-constexpr size_t N{8};
+constexpr int N{8};
 constexpr double r_e{0.08};
 constexpr double r_i{0.04};
 constexpr double base_length{0.8};
 constexpr double sqrt2{1.4142};
 
 void mesh_arrow(std::vector<double> &vertices, std::vector<unsigned int> &faces,
-                size_t n_angles) {
+                int n_angles) {
   /* Returns a mesh representing an arrow of coordinates direction (1, 0, 0),
    * with base at (0,0,0). */
 
-  size_t n_vertices = 3 * n_angles + 2;
-  size_t n_faces = 6 * n_angles;
+  int n_vertices = 3 * n_angles + 2;
+  int n_faces = 6 * n_angles;
 
   vertices.resize(n_vertices * 3, r_i);
 
@@ -26,7 +26,7 @@ void mesh_arrow(std::vector<double> &vertices, std::vector<unsigned int> &faces,
   std::vector<double> theta_ext(n_angles);
   std::vector<double> theta_int(n_angles);
 
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     theta_ext.at(i) = 2 * pi / n_angles * i + pi / n_angles;
     theta_int.at(i) = 2 * pi / n_angles * i;
   }
@@ -34,18 +34,18 @@ void mesh_arrow(std::vector<double> &vertices, std::vector<unsigned int> &faces,
   vertices[1] = 0;
   vertices[2] = 0;
   // shaft base
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     vertices[3 * (i + 1) + 1] = r_i * sin(theta_int.at(i));
     vertices[3 * (i + 1) + 2] = r_i * cos(theta_int.at(i));
   }
   // cone junction
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     vertices[3 * (i + n_angles + 1)] = base_length;
     vertices[3 * (i + n_angles + 1) + 1] = r_i * sin(theta_int.at(i));
     vertices[3 * (i + n_angles + 1) + 2] = r_i * cos(theta_int.at(i));
   }
   // cone base
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     vertices[3 * (i + 2 * n_angles + 1)] = base_length;
     vertices[3 * (i + 2 * n_angles + 1) + 1] = r_e * sin(theta_ext.at(i));
     vertices[3 * (i + 2 * n_angles + 1) + 2] = r_e * cos(theta_ext.at(i));
@@ -56,38 +56,38 @@ void mesh_arrow(std::vector<double> &vertices, std::vector<unsigned int> &faces,
   vertices.at(3 * (3 * n_angles + 1) + 2) = 0;
 
   // shaft base
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces.at(3 * i) = (i + 1) % n_angles + 1;
     faces.at(3 * i + 1) = i + 1;
     faces.at(3 * i + 2) = 0;
   }
 
   // shaft
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces[3 * (i + n_angles)] = i + 1;
     faces[3 * (i + n_angles) + 1] = (i + 1) % n_angles + 1;
     faces[3 * (i + n_angles) + 2] = n_angles + i + 1;
   }
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces[3 * (i + 2 * n_angles)] = (i + 1) % n_angles + 1;
     faces[3 * (i + 2 * n_angles) + 1] = n_angles + (i + 1) % n_angles + 1;
     faces[3 * (i + 2 * n_angles) + 2] = n_angles + i + 1;
   }
 
   // Cone base
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces[3 * (i + 3 * n_angles)] = n_angles + i % n_angles + 1;
     faces[3 * (i + 3 * n_angles) + 1] = n_angles + (i + 1) % n_angles + 1;
     faces[3 * (i + 3 * n_angles) + 2] = 2 * n_angles + i + 1;
   }
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces[3 * (i + 4 * n_angles)] = n_angles + (i + 1) % n_angles + 1;
     faces[3 * (i + 4 * n_angles) + 1] = 2 * n_angles + (i + 1) % n_angles + 1;
     faces[3 * (i + 4 * n_angles) + 2] = 2 * n_angles + i + 1;
   }
 
   // Cone tip
-  for (size_t i = 0; i < n_angles; ++i) {
+  for (int i = 0; i < n_angles; ++i) {
     faces.at(3 * (i + 5 * n_angles)) = 2 * n_angles + i + 1;
     faces.at(3 * (i + 5 * n_angles) + 1) =
         2 * n_angles + (i + 1) % n_angles + 1;
@@ -139,8 +139,8 @@ void colored_cross(std::vector<double> &vertices,
 
   mesh_arrow(vertices, faces, N);
 
-  size_t n_vertices = vertices.size() / 3;
-  size_t n_faces = faces.size() / 3;
+  int n_vertices = (int)vertices.size() / 3;
+  int n_faces = (int)faces.size() / 3;
   vertices.resize(vertices.size() * 3);
   faces.resize(faces.size() * 3);
 
@@ -161,7 +161,7 @@ void colored_cross(std::vector<double> &vertices,
 
   std::vector<Quaternion> vert_quat(n_vertices);
 
-  for (size_t i = 0; i < n_vertices; ++i) {
+  for (int i = 0; i < n_vertices; ++i) {
     vert_quat.at(i)[0] = 0;
     vert_quat.at(i)[1] = vertices[3 * i];
     vert_quat.at(i)[2] = vertices[3 * i + 1];
@@ -169,7 +169,7 @@ void colored_cross(std::vector<double> &vertices,
   }
 
   // y arrow
-  for (size_t i = 0; i < n_vertices; ++i) {
+  for (int i = 0; i < n_vertices; ++i) {
     vert_tmp.multiply(qx_y, vert_quat.at(i));
     vert_tmp = vert_tmp * qx_y_inv;
     vertices.at((i + n_vertices) * 3) = vert_tmp[1];
@@ -178,7 +178,7 @@ void colored_cross(std::vector<double> &vertices,
   }
 
   // z arrow
-  for (size_t i = 0; i < n_vertices; ++i) {
+  for (int i = 0; i < n_vertices; ++i) {
     vert_tmp.multiply(qx_z, vert_quat.at(i));
     vert_tmp = vert_tmp * qx_z_inv;
     vertices.at((i + 2 * n_vertices) * 3) = vert_tmp[1];
@@ -187,14 +187,14 @@ void colored_cross(std::vector<double> &vertices,
   }
 
   // Add offset to the faces indices
-  for (size_t i = 0; i < n_faces * 3; ++i) {
+  for (int i = 0; i < n_faces * 3; ++i) {
     faces.at(i + n_faces * 3) = n_vertices + faces.at(i);
     faces.at(i + n_faces * 6) = 2 * n_vertices + faces.at(i);
   }
 
   colors.resize(vertices.size(), 0);
 
-  for (size_t i = 0; i < n_vertices; ++i) {
+  for (int i = 0; i < n_vertices; ++i) {
     colors.at(i * 3) = 1;                        // red
     colors.at((i + n_vertices) * 3 + 1) = 1;     // blue
     colors.at((i + 2 * n_vertices) * 3 + 2) = 1; // green
@@ -212,7 +212,7 @@ void MeshRender::set_axis_cross() {
 
   std::vector<double> axis_colors;
   colored_cross(axis_vertices, axis_faces, axis_colors);
-  size_t n_vertices = axis_vertices.size() / 3;
+  int n_vertices = (int)axis_vertices.size() / 3;
 
   // Adding a cube at the origin
   std::vector<double> cube_vertices;
